rename types and files in runner
