+++ servers/execution/runner/src/interfaces/command.interface.ts
